requirements.txt
anthropic
chromadb
flask
flask-cors
gunicorn
python-dotenv
requests